refactor: fe - component layout 묶기 #146
