+++ frontend/src/pages/workspace/header/index.tsx
@@ -1,20 +1,12 @@
-import Modal from '@components/modal';
-import useModal from '@hooks/useModal';
 import { useState, useEffect } from 'react';
 import LeftSide from '../left-side';
 import RightSide from '../right-side';
-import ShareModal from '../share-modal';
 import { Container } from './index.style';
 import { Workspace } from '@api/workspace';
+import { HeaderProps } from './index.type';
 
-interface HeaderProps {
-	workspaceId: string;
-}
-
-function Header({ workspaceId }: HeaderProps) {
+function Header({ workspaceId, openShareModal }: HeaderProps) {
 	const [workspace, setWorkspaceName] = useState<string>('');
-	const { isOpenModal, openModal, modalRef, closeModal } = useModal();
-	const [modalContent, setModalContent] = useState(<></>);
 
 	useEffect(() => {
 		async function getMetaData() {
@@ -24,23 +16,12 @@ function Header({ workspaceId }: HeaderProps) {
 		getMetaData();
 	}, []);
 
-	const openShareModal = () => {
-		openModal();
-		setModalContent(<ShareModal modalRef={modalRef} id={workspaceId} closeModal={closeModal} />);
-	};
-
 	return (
-		<>
-			<Container>
-				<LeftSide />
-				<p className="title">{workspace}</p>
-				<RightSide openShareModal={openShareModal} />
-			</Container>
-
-			<Modal isOpen={isOpenModal} modalRef={modalRef} width={600} height={400}>
-				{modalContent}
-			</Modal>
-		</>
+		<Container>
+			<LeftSide />
+			<p className="title">{workspace}</p>
+			<RightSide openShareModal={openShareModal} />
+		</Container>
 	);
 }
