+++ frontend/src/pages/workspace/header/index.type.ts
@@ -1,0 +1,4 @@
+export interface HeaderProps {
+	workspaceId: string;
+	openShareModal: () => void;
+}
