+++ frontend/src/pages/workspace/layout/index.style.tsx
@@ -1,0 +1,3 @@
+import styled from 'styled-components';
+
+export const Container = styled.div``;
